requirements.txt

# Kasparro V2 Production Requirements
# Python >= 3.10

# Core Data Science Stack
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
statsmodels==0.14.1

# Utilities / Config
PyYAML==6.0.1
tenacity==8.2.3

# Logging / Time handling
python-dateutil==2.8.2
pytz==2023.3
numexpr==2.8.7

# Optional (used by OpenAI creative generator)
openai==1.12.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
